feat: enemy movement still janky needs refinement

+++ Assets/Scripts/New folder/EnemyDefinition.cs
@@ -6,12 +6,12 @@ public class EnemyDefinition : ScriptableObject, IUnitDefinition
 {
     [SerializeField] private GameObject enemyPrefab;
     [SerializeField] private int health;
-    [SerializeField] private int speed;
+    [SerializeField] private float speed;
     [SerializeField] private string description;
 
     public GameObject EnemyPrefab { get { return enemyPrefab; } }
     public int Health { get { return health; } }
-    public int Speed { get { return speed; } }
+    public float Speed { get { return speed; } }
     public string Description { get { return description; } }
 
     public event Action OnRebuild;
